Added score view and logic

+++ onlinecourse/models.py
@@ -96,7 +96,7 @@ class Enrollment(models.Model):
 
 class Question(models.Model):
     question_text = models.CharField(max_length=256)
-    grade_point = models.FloatField()
+    grade_point = models.IntegerField()
     course = models.ForeignKey(Course, on_delete=models.CASCADE)
 
     def __str__(self):
